+++ library/ui/src/main/java/com/google/android/exoplayer2/ui/PlayerControlView.java
@@ -41,10 +41,13 @@ import android.view.LayoutInflater;
 import android.view.MotionEvent;
 import android.view.View;
 import android.view.ViewGroup;
+import android.view.accessibility.AccessibilityEvent;
 import android.widget.FrameLayout;
 import android.widget.ImageView;
 import android.widget.TextView;
+import androidx.annotation.DoNotInline;
 import androidx.annotation.Nullable;
+import androidx.annotation.RequiresApi;
 import com.google.android.exoplayer2.C;
 import com.google.android.exoplayer2.ControlDispatcher;
 import com.google.android.exoplayer2.ExoPlayerLibraryInfo;
@@ -339,6 +342,8 @@ public class PlayerControlView extends FrameLayout {
   private long[] extraAdGroupTimesMs;
   private boolean[] extraPlayedAdGroups;
   private long currentWindowOffset;
+  private long currentPosition;
+  private long currentBufferedPosition;
 
   public PlayerControlView(Context context) {
     this(context, /* attrs= */ null);
@@ -783,6 +788,7 @@ public class PlayerControlView extends FrameLayout {
       }
       updateAll();
       requestPlayPauseFocus();
+      requestPlayPauseAccessibilityFocus();
     }
     // Call hideAfterTimeout even if already visible to reset the timeout.
     hideAfterTimeout();
@@ -831,17 +837,29 @@ public class PlayerControlView extends FrameLayout {
       return;
     }
     boolean requestPlayPauseFocus = false;
+    boolean requestPlayPauseAccessibilityFocus = false;
     boolean shouldShowPauseButton = shouldShowPauseButton();
     if (playButton != null) {
       requestPlayPauseFocus |= shouldShowPauseButton && playButton.isFocused();
+      requestPlayPauseAccessibilityFocus |=
+          Util.SDK_INT < 21
+              ? requestPlayPauseFocus
+              : (shouldShowPauseButton && Api21.isAccessibilityFocused(playButton));
       playButton.setVisibility(shouldShowPauseButton ? GONE : VISIBLE);
     }
     if (pauseButton != null) {
       requestPlayPauseFocus |= !shouldShowPauseButton && pauseButton.isFocused();
+      requestPlayPauseAccessibilityFocus |=
+          Util.SDK_INT < 21
+              ? requestPlayPauseFocus
+              : (!shouldShowPauseButton && Api21.isAccessibilityFocused(pauseButton));
       pauseButton.setVisibility(shouldShowPauseButton ? VISIBLE : GONE);
     }
     if (requestPlayPauseFocus) {
       requestPlayPauseFocus();
+    }
+    if (requestPlayPauseAccessibilityFocus) {
+      requestPlayPauseAccessibilityFocus();
     }
   }
 
@@ -1021,14 +1039,21 @@ public class PlayerControlView extends FrameLayout {
       position = currentWindowOffset + player.getContentPosition();
       bufferedPosition = currentWindowOffset + player.getContentBufferedPosition();
     }
-    if (positionView != null && !scrubbing) {
+    boolean positionChanged = position != currentPosition;
+    boolean bufferedPositionChanged = bufferedPosition != currentBufferedPosition;
+    currentPosition = position;
+    currentBufferedPosition = bufferedPosition;
+
+    // Only update the TextView if the position has changed, else TalkBack will repeatedly read the
+    // same position to the user.
+    if (positionView != null && !scrubbing && positionChanged) {
       positionView.setText(Util.getStringForTime(formatBuilder, formatter, position));
     }
     if (timeBar != null) {
       timeBar.setPosition(position);
       timeBar.setBufferedPosition(bufferedPosition);
     }
-    if (progressUpdateListener != null) {
+    if (progressUpdateListener != null && (positionChanged || bufferedPositionChanged)) {
       progressUpdateListener.onProgressUpdate(position, bufferedPosition);
     }
 
@@ -1062,6 +1087,15 @@ public class PlayerControlView extends FrameLayout {
       playButton.requestFocus();
     } else if (shouldShowPauseButton && pauseButton != null) {
       pauseButton.requestFocus();
+    }
+  }
+
+  private void requestPlayPauseAccessibilityFocus() {
+    boolean shouldShowPauseButton = shouldShowPauseButton();
+    if (!shouldShowPauseButton && playButton != null) {
+      playButton.sendAccessibilityEvent(AccessibilityEvent.TYPE_VIEW_FOCUSED);
+    } else if (shouldShowPauseButton && pauseButton != null) {
+      pauseButton.sendAccessibilityEvent(AccessibilityEvent.TYPE_VIEW_FOCUSED);
     }
   }
 
@@ -1337,6 +1371,14 @@ public class PlayerControlView extends FrameLayout {
       } else if (shuffleButton == view) {
         controlDispatcher.dispatchSetShuffleModeEnabled(player, !player.getShuffleModeEnabled());
       }
+    }
+  }
+
+  @RequiresApi(21)
+  private static final class Api21 {
+    @DoNotInline
+    public static boolean isAccessibilityFocused(View view) {
+      return view.isAccessibilityFocused();
     }
   }
 }